feat: New Profil UI
